Added formatter to manage board formatting for displaying to the console.
