+++ lib/player.ex
@@ -1,0 +1,6 @@
+defmodule Player do
+  def get_move(marker) do
+    Messages.get(:get_move, marker)
+    |> Console.get_input()
+  end
+end
