+++ test/console_test.exs
@@ -14,47 +14,6 @@ defmodule ConsoleTest do
              "You have started a new game of TTT!\n\nHOW TO PLAY\n===========\nPlayers alternate placing different markers on the board until either one player has three in a row, horizontally, vertically, or diagonally; or all nine squares are filled.\nIf a player is able to draw three of their markers in a row, then that player wins.\n\n"
   end
 
-  test "format empty cell adds a pipe and spaces", do: assert(Console.format_cell(1) == "| 1 ")
-
-  test "format filled cell adds a pipe and spaces", do: assert(Console.format_cell("X") == "| X ")
-
-  test "format_board returns a list of formatted cells" do
-    assert Console.format_board([1, "X"]) == ["| 1 ", "| X "]
-  end
-
-  test "split_into_rows takes a list of 9 elements returns a chunked list of 3 rows" do
-    assert Console.split_into_rows([1, "X", 3, 4, 5, 6, "O", "X", 9]) == [
-             [1, "X", 3],
-             [4, 5, 6],
-             ["O", "X", 9]
-           ]
-  end
-
-  test "split_into_rows takes a list of 16 elements returns a chunked list of 4 rows" do
-    assert Console.split_into_rows([1, 2, 3, 4, 5, 6, 6, 8, 9, 10, 11, 12, 13, 14, 15, 16]) == [
-             [1, 2, 3, 4],
-             [5, 6, 6, 8],
-             [9, 10, 11, 12],
-             [13, 14, 15, 16]
-           ]
-  end
-
-  test "format_rows inserts a divider in chunked board" do
-    assert Console.add_row_dividers([
-             ["| 1 ", "| X ", "| 3 "],
-             ["| 4 ", "| 5 ", "| 6 "],
-             ["| O ", "| X ", "| 9 "]
-           ]) ==
-             [
-               " \n  ———————————\n ",
-               ["| 1 ", "| X ", "| 3 "],
-               "|\n  ———————————\n ",
-               ["| 4 ", "| 5 ", "| 6 "],
-               "|\n  ———————————\n ",
-               ["| O ", "| X ", "| 9 "],
-               "|\n  ———————————\n "
-             ]
-  end
 
   test "user can see a board of 1 - 9" do
     response = fn ->
